pyproject.toml
[tool.poetry]
name = "sports_line_scrape"
version = "0.1.2"
description = "Package used to scrape odds at sportsline.com"
authors = ["Mason Wrolstad <mason.wrolstad@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.2"
python-dateutil = "^2.8.2"
pytz = "^2022.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
